ValueObjectを整理
