+++ RegisterOwnCompanyCalendarApplication/RegisterOwnCompanyCalendarUseCase.cs
@@ -1,0 +1,7 @@
+﻿namespace RegisterOwnCompanyCalendarApplication
+{
+    public class RegisterOwnCompanyCalendarUseCase
+    {
+
+    }
+}
