+++ VerifyAttendanceCSV.sln
@@ -27,6 +27,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DetermineDifferenceApplicat
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wada.AttendanceCSVTests", "Wada.AttendanceCSVTests\Wada.AttendanceCSVTests.csproj", "{D3B284C4-5E49-4109-AC02-3440FC9B7A1F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RegisterOwnCompanyCalendarApplication", "RegisterOwnCompanyCalendarApplication\RegisterOwnCompanyCalendarApplication.csproj", "{9297CA21-4E57-44E1-888E-CAB442E855CB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -65,6 +67,10 @@ Global
 		{D3B284C4-5E49-4109-AC02-3440FC9B7A1F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D3B284C4-5E49-4109-AC02-3440FC9B7A1F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D3B284C4-5E49-4109-AC02-3440FC9B7A1F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9297CA21-4E57-44E1-888E-CAB442E855CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9297CA21-4E57-44E1-888E-CAB442E855CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9297CA21-4E57-44E1-888E-CAB442E855CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9297CA21-4E57-44E1-888E-CAB442E855CB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -77,6 +83,7 @@ Global
 		{F8C0B4C3-2769-4A8F-8E74-A5BE6811CA83} = {CA9E2383-0E27-4018-9723-183AC5FC2F18}
 		{CEF632F3-95F2-48A2-9436-3387E0A8F050} = {F09417BC-393B-4DF0-9DE4-02604E8136E4}
 		{D3B284C4-5E49-4109-AC02-3440FC9B7A1F} = {CA9E2383-0E27-4018-9723-183AC5FC2F18}
+		{9297CA21-4E57-44E1-888E-CAB442E855CB} = {F09417BC-393B-4DF0-9DE4-02604E8136E4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A6BBC0BD-A5F6-44C4-B7A6-7FAD6289890C}
