+++ Wada.AttendanceTableService/IWadaHolidayRepository.cs
@@ -1,4 +1,4 @@
-﻿using Wada.AttendanceTableService.AttendanceTableAggregation;
+﻿using Wada.AttendanceTableService.ValueObjects;
 
 namespace Wada.AttendanceTableService
 {
